refactor: move delete category logic to CategoryDetailsViewModel
